Freq_Stats: Show Toast if stats aren't persistent
This should show a toast to notify the user when a core is offline that the stats will be inaccurate if the stats path is not the persistent stats path.

The shortcoming here is that it only accounts for a core BEING offline when the fragment's view is generated. It isn't checking for the core being plugged/unplugged outside of that.